+++ config/rest-presenter.php
@@ -8,6 +8,7 @@ use XtendPackages\RESTPresenter\Middleware\VerifyApiKey;
 use XtendPackages\RESTPresenter\Resources\Users\Presenters;
 
 return [
+    'scan_paths' => explode(',', env('REST_PRESENTER_SCAN_PATHS', app_path().':App')),
     'generator' => [
         'path' => env('REST_PRESENTER_GENERATOR_PATH', 'app/Api'),
         'namespace' => env('REST_PRESENTER_GENERATOR_NAMESPACE', 'App\Api'),
@@ -45,7 +46,7 @@ return [
         ],
         'logout_revoke_all_tokens' => env('REST_PRESENTER_AUTH_LOGOUT_REVOKE_ALL_TOKENS', false),
         'rate_limit' => [
-            'max_attempts' => (int) env('REST_PRESENTER_AUTH_RATE_LIMIT_MAX_ATTEMPTS', 5),
+            'max_attempts' => (int) env('REST_PRESENTER_AUTH_RATE_LIMIT_MAX_ATTEMPTS', 5), // @phpstan-ignore-line
         ],
     ],
     'exporters' => [