+++ src/Concerns/InteractsWithConfig.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace XtendPackages\RESTPresenter\Concerns;
+
+trait InteractsWithConfig
+{
+    // @phpstan-ignore-next-line
+    protected function config(string $key, mixed $default = null)
+    {
+        return config($key, $default) ?? $default;
+    }
+}