+++ src/Models/User.php
@@ -16,6 +16,8 @@ use XtendPackages\RESTPresenter\Factories\UserFactory;
  * @property string $password
  *
  * @mixin \Illuminate\Database\Eloquent\Builder<User>
+ *
+ * @use HasFactory<UserFactory>
  */
 class User extends Authenticatable
 {
@@ -44,6 +46,9 @@ class User extends Authenticatable
         'remember_token',
     ];
 
+    /**
+     * The model factory for this model.
+     */
     protected static function newFactory(): UserFactory
     {
         return UserFactory::new();